refactor: 페이지 단위로 컴포넌트 폴더 분리
