+++ src/components/DisasterNotiSettings/index.js
@@ -1,0 +1,2 @@
+export { default as SelectNaturalDisasterBottomSheet } from './SelectNaturalDisaster/SelectNaturalDisasterBottomSheet';
+export { default as SelectSocialDisasterBottomSheet } from './SelectSocialDisaster/SelectSocialDisasterBottomSheet';
