+++ src/components/Settings/index.js
@@ -1,0 +1,1 @@
+export { default as AddressBottomSheet } from './NotificationContent/NotificationContentBottomSheet';
